Remove enkf.py from dark_storage
