+++ src/ert/dark_storage/endpoints/experiments.py
@@ -3,7 +3,7 @@ from uuid import UUID
 from fastapi import APIRouter, Body, Depends
 
 from ert.dark_storage import json_schema as js
-from ert.dark_storage.enkf import get_storage
+from ert.dark_storage.common import get_storage
 from ert.shared.storage.extraction import create_priors
 from ert.storage import Storage
